ingestion_service/src/api/errors.py
from fastapi import Request
from fastapi.responses import JSONResponse
from fastapi.exceptions import RequestValidationError
from starlette import status

from src.api.v1.models import ErrorResponse


def register_error_handlers(app) -> None:
    @app.exception_handler(RequestValidationError)
    async def validation_error_handler(request: Request, exc: RequestValidationError):
        error = ErrorResponse(
            error_code="INVALID_REQUEST",
            message="Request validation failed",
            details={"errors": exc.errors()},
        )
        return JSONResponse(
            status_code=status.HTTP_400_BAD_REQUEST,
            content=error.model_dump(),
        )

    @app.exception_handler(Exception)
    async def unhandled_exception_handler(request: Request, exc: Exception):
        error = ErrorResponse(
            error_code="INTERNAL_ERROR",
            message="Internal server error",
        )
        return JSONResponse(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            content=error.model_dump(),
        )
